🤖 refactor: 개인봉사자 defaultValue 바인딩 useEffect로 변경
